Normalize the scale for axes
This PR take the highest count from positive and negative
votes and uses that as the max value for each axis.

This change provides a more accurate representation of
sentiment for a page, but it makes it harder to identify
pages with negative counts that we might want
to act on despite the page's even higher positive counts.

Fixes #3.